fix(build): extract detectPaymentFrequency from 'use server' module (Next.js 15)
Next.js 15 requires every export in a 'use server' file to be async.
detectPaymentFrequency is a pure synchronous utility; move it to
src/lib/dividends/payment-frequency.ts so RSC compilation succeeds.
