+++ apps/frontend/src/app/dividends/__tests__/dividend-positions.test.ts
@@ -29,8 +29,8 @@ vi.mock('@/app/trading/actions', () => ({
 import {
   getDividendPositions,
   getDividendSummary,
-  detectPaymentFrequency,
 } from '../actions';
+import { detectPaymentFrequency } from '@/lib/dividends/payment-frequency';
 import { createClient } from '@/lib/supabase/server';
 import { getStockPositions } from '@/app/trading/actions';
