+++ apps/frontend/src/app/dividends/actions.ts
@@ -8,6 +8,7 @@ import type {
   DividendSummaryResult,
   PaymentFrequency,
 } from '@/types/dividends';
+import { detectPaymentFrequency } from '@/lib/dividends/payment-frequency';
 
 // Consumers should import DividendPosition, DividendSummaryResult, PaymentFrequency directly from '@/types/dividends'.
 
@@ -909,37 +910,6 @@ export async function saveDividendEstimations(
 // Source-of-truth: stock_positions (via getStockPositions) enriched with
 // TTM yield from dividend_payments and forward yield from dividend_accruals.
 // dividend_ticker_data is NOT used here (empty table).
-
-/**
- * Detects dividend payment frequency from a list of ex-dates.
- * Returns null when there are fewer than 2 dates to compare.
- *
- * Thresholds (average interval in days):
- *   ≤ 40   → monthly
- *   ≤ 100  → quarterly
- *   ≤ 200  → semi-annual
- *   ≤ 450  → annual
- *   > 450  → irregular
- */
-export function detectPaymentFrequency(dates: string[]): PaymentFrequency {
-  if (dates.length < 2) return dates.length === 1 ? 'annual' : null;
-
-  const sorted = [...dates].sort();
-  let totalDays = 0;
-  for (let i = 1; i < sorted.length; i++) {
-    const diff =
-      (new Date(sorted[i]).getTime() - new Date(sorted[i - 1]).getTime()) /
-      (1000 * 60 * 60 * 24);
-    totalDays += diff;
-  }
-  const avgDays = totalDays / (sorted.length - 1);
-
-  if (avgDays <= 40) return 'monthly';
-  if (avgDays <= 100) return 'quarterly';
-  if (avgDays <= 200) return 'semi-annual';
-  if (avgDays <= 450) return 'annual';
-  return 'irregular';
-}
 
 function paymentsPerYear(freq: PaymentFrequency): number {
   switch (freq) {
